+++ backend/src/baserow/api/groups/users/serializers.py
@@ -39,22 +39,36 @@ class GroupUserSerializer(serializers.ModelSerializer):
         return object.user.email
 
 
-class GroupUserGroupSerializer(serializers.ModelSerializer):
+class GroupUserGroupSerializer(serializers.Serializer):
     """
-    This serializers returns all the fields that the GroupSerializer has, but also
-    some user specific values related to the group user relation.
+    This serializers includes relevant fields of the Group model, but also
+    some GroupUser specific fields related to the group user relation.
+
+    Additionally, the list of users are included for each group.
     """
 
-    class Meta:
-        model = GroupUser
-        fields = ("order", "permissions")
+    # Group fields
+    id = serializers.IntegerField(
+        source="group.id", read_only=True, help_text="Group id."
+    )
+    name = serializers.CharField(
+        source="group.name", read_only=True, help_text="Group name."
+    )
+    users = GroupUserSerializer(
+        many=True,
+        source="group.groupuser_set",
+        required=False,
+        read_only=True,
+        help_text="List of all group users.",
+    )
 
-    def to_representation(self, instance):
-        from baserow.api.groups.serializers import GroupSerializer
-
-        data = super().to_representation(instance)
-        data.update(GroupSerializer(instance.group).data)
-        return data
+    # GroupUser fields
+    order = serializers.IntegerField(
+        read_only=True, help_text="The requesting user's order within the group users."
+    )
+    permissions = serializers.CharField(
+        read_only=True, help_text="The requesting user's permissions for the group."
+    )
 
 
 class UpdateGroupUserSerializer(serializers.ModelSerializer):